requirements.txt
# LiveKit Agents SDK
livekit-agents==1.4.2

# STT Provider
livekit-plugins-deepgram==1.4.2

# LLM Provider
livekit-plugins-openai==1.4.2

# TTS Provider
livekit-plugins-elevenlabs==1.4.2

# VAD (Voice Activity Detection)
livekit-plugins-silero==1.4.2

# Environment variables
python-dotenv==1.2.1

# Async HTTP
aiohttp==3.13.3

# API server
fastapi==0.115.12
uvicorn==0.34.0

# Database
psycopg2-binary
